drink_db/requirements.txt
git+https://gitlab.com/jfautley/alexapy.git
python-dotenv
tornado==6.0.3